section 4.27 Specular Lights

+++ Light.cpp
@@ -17,8 +17,8 @@ Light::Light(GLfloat red, GLfloat green, GLfloat blue,
 {
 }
 
-void Light::useLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-                     GLfloat diffuseIntensityLocation, GLfloat directionLocation)
+void Light::useLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+                     GLuint diffuseIntensityLocation, GLuint directionLocation)
 {
     glUniform3f(ambientColorLocation, color.x, color.y, color.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
