+++ Light.hpp
@@ -17,7 +17,7 @@ public:
           GLfloat aIntensity,
           GLfloat xDir, GLfloat yDir, GLfloat zDir,
           GLfloat dIntensity);
-    void useLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-                  GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+    void useLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+                  GLuint diffuseIntensityLocation, GLuint directionLocation);
     ~Light();
 };
